Add zsh completion
Wonderfully, zsh can interpret bash completion scripts, so this came quite easy.

This also includes a fix for a semi-bug zsh uncovered. Previously I was pattern
matching [a-z]. In bash this is case-insensitive, but in zsh it's case
sensitive. I'm surprised it's insensitive in bash, but in any case, it's fixed
now and works in both.

Works towards issue #51