Replace non-POSIX `readlink` with shell
It turns out that `readlink` is non-POSIX and very non-portable. Solaris (at
least Solaris 10) does not have `readlink`, and OS X does not have
`readlink -f` (which we used).

This approach breaks the rule of "never parse the output of `ls`". However,
this is the exception that proves the rule, as the POSIX docs are very specific
about how `ls` should format the output when listing a symlink.
http://pubs.opengroup.org/onlinepubs/007904975/utilities/ls.html#tag_04_81_10

I tested this on Debian, FreeBSD, Solaris 10, and OS X. All conformed to the
spec.

This method will fail if:
 - the symlink name or dirname contains '->' (which is uncommon IMO).
 - the parent dir is a symlink to the folder which contains zfsnap.sh

These are unlikely scenarios in my opinion, and the $ZFSNAP_LIB_DIR variable is
available in case anyone has a less-common setup. Handling these corner cases is
possible, but I saw no reason to inflate the code unnecessarily to make this
approach canonical. Others are invited to do so if they're so inclined.

Also a `dirname` call is removed.