Add IsInt function rather than [ $var -ge 0 2>/dev/null ]
posh would segfault using the following check:
[ "$var" -ge 0 2>/dev/null ] || return 1

This Pattern Matching method works without segfault
[ -z "${var##*[!0-9]*}" ] && return 1

I was curious about performance, so I ran some quick tests. And the results are
surprising (all run under dash 0.5.7).

1) original (average 4.14 seconds)
  for i in `seq 1 1000000`; do
    [ "$i" -ge 0 2> /dev/null ] || echo "is not int"
  done

2) shell pattern (average 3.11 seconds)
  for i in `seq 1 1000000`; do
    [ -z "${i##*[!0-9]*}" ] && echo "is not int"
  done

WTF. The shell pattern method is faster. But why? My first guess was that
redirection to null was the culprit.

3) original w/o redirection (average 3.81 seconds)
  for i in `seq 1 1000000`; do
    [ "$i" -ge 0 ] || echo "is not int"
  done

Faster than before, but /still/ slower than pattern matching. On a hunch, I
removed the quotes around the variable (bad idea for safety).

4) original w/o redirection and w/o quoting (average 2.98 seconds)
  for i in `seq 1 1000000`; do
    [ $i -ge 0 ] || echo "is not int"
  done

So it seems quoting costs a lot. Perhaps there's some conversion from int to
string back into int that hurts here. I'm not sure though.

Then I thought about the quotes on the pattern matching...

5) shell pattern w/o quotes (average 2.78 seconds)
  for i in `seq 1 1000000`; do
    [ -z ${i##*[!0-9]*} ] && echo "is not int"
  done

So, pattern matching is faster, when all else is equal. I don't know why, but I
plan to ask the dash guys and see if they have an idea.

One last benchmark:

6) shell pattern with function (average 4.62 seconds)
  IsInt() {
    [ -z "${1##*[!0-9]*}" ] && return 1 || return 0
  }

  for i in `seq 1 1000000`; do
    IsInt "$i" || echo "is not int"
  done

The speed advantage of pattern matching (in this scenario at least) gets lost
when accesing via a function, which is to be expected. In the end, it's worth it
for readability and testing. When all is said and done, this check isn't /that/
much slower than the original method that crashed on posh.

A bit of overkill for a few shell scripts, but I was curious, and the results
are interesting.