Add test_vagrants.sh script for cross-OS testing
"cd vagrant/ ; ./test_vagrants.sh --all" will start and the stop each of the
Vagrant boxes.

The actual copying and executing happens in the VagrantFile, via the
provisioning which runs each time a box is started. Given there are so many
boxes, they are not kept up, as together they'd consume quite a bit of RAM.

Works towards #52