Fix acquisition of ZFSNAP_LIB_DIR
Use /bin/ls rather than ls to avoid aliases.

More importantly, simplified the stripping of ls's symlink output. The old
method didn't work on my home server but /does/ work on a server at work -- (for
reasons which still elude me). The only advantage of the old method was that it
would remove more than one leading space, but POSIX's documents show there
shall be only one space.

    "%s -> %s", <pathname of link>, <contents of link>

This way is simpler and better (in that it works ;-).